Refactor providers to core instead of swing
