+++ core/src/main/java/info/openrocket/core/startup/jij/ClasspathProvider.java
@@ -1,4 +1,4 @@
-package info.openrocket.swing.startup.jij;
+package info.openrocket.core.startup.jij;
 
 import java.net.URL;
 import java.util.List;
