+++ core/src/main/java/info/openrocket/core/startup/jij/ConfigurableStreamHandlerFactory.java
@@ -1,4 +1,4 @@
-package info.openrocket.swing.startup.jij;
+package info.openrocket.core.startup.jij;
 
 import java.net.URLStreamHandler;
 import java.net.URLStreamHandlerFactory;
