+++ core/src/main/java/info/openrocket/core/startup/jij/PluginClasspathProvider.java
@@ -1,4 +1,4 @@
-package info.openrocket.swing.startup.jij;
+package info.openrocket.core.startup.jij;
 
 import java.io.File;
 import java.net.MalformedURLException;
